fix: harden temporal rollup review concerns
